Add feature flag for transcripts
The `Settings.enabled_features.transcripts` flag was added to control
whether the viewer will display a track element containing the
transcript for a video.

The setting can be forced on by adding `?trascripts=true` to the URL for
the embed viewer. This will require you to look at the view directly
instead of embedded in PURL or SearchWorks.

Resolves #1497