Add headers to companion window components
Connects to #1631

This commit adds headers to both the left and right companion window components.
It also extracts the rights component out of the metadata list component. Note
that this commit only affects the HTML DOM and JavaScript; the CSS to style the
headers similar to Mirador is not included.