Improve local development
This commit aims to improve local development of the media player in particular, by running a Stacks container out of a sibling directory and by serving files without requiring a media server (such as Wowza).

Includes the following changes:

* Add a rake task to pull downloadable files from prod stacks into the local fake stacks root
* Ignore the fake stacks root (in `public/file/`)
* Use the druid-tools gem to construct the directory structure Stacks expects to serve files from (only in development bundler group)
* Pull in stacks' docker-compose so we can run it as part of sul-embed's `bin/dev` and mount the fake stacks root as a volume in the container
* Update README
* Add development configuration as a file

NOTE: Depends on https://github.com/sul-dlss/stacks/pull/1053, and on having an up-to-date checkout of stacks in a sibling directory. We can get rid of this in the future if we publish the stacks image.