Intercept links from IIIF manifest to set target and rels
Fixes #1963

Links from the IIIF manifest, such as for "Available Online" in the metadata panel, are generated in PURL without targets or rels, and as such, these links wind up opening inside the embed iframe leading to a dizzying link-ception scenario as documented in the associated issue. This commit fixes that behavior in the same way Mirador already does: https://github.com/ProjectMirador/mirador/blob/cefe7b7b520ae1cb637d935c7b3cfceaff96b14a/src/components/SanitizedHtml.js#L17-L25 Namely, by sniffing each IIIF manifest value to see if it is a node with a target, such as an anchor element, and coercing the target to "_blank" and setting the expected rels. This forces links to open in a new window.